refactor: `NamespaceAccessor` typing/protocol changes
Everything **but** the runtime change to `requires._unwrap_context`

https://github.com/narwhals-dev/narwhals/pull/3127#discussion_r2346948601
