+++ narwhals/_compliant/any_namespace.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, ClassVar, Protocol
 
-from narwhals._utils import CompliantT_co, _StoresCompliant
+from narwhals._utils import CompliantT_co, Implementation, _StoresCompliant
 
 if TYPE_CHECKING:
     from typing import Callable
@@ -26,6 +26,14 @@ __all__ = [
 class NamespaceAccessor(_StoresCompliant[CompliantT_co], Protocol[CompliantT_co]):
     _accessor: ClassVar[Accessor]
 
+    @property
+    def implementation(self) -> Implementation:
+        return self.compliant._implementation
+
+    @property
+    def backend_version(self) -> tuple[int, ...]:
+        return self.implementation._backend_version()
+
 
 class CatNamespace(NamespaceAccessor[CompliantT_co], Protocol[CompliantT_co]):
     _accessor: ClassVar[Accessor] = "cat"
@@ -33,7 +41,7 @@ class CatNamespace(NamespaceAccessor[CompliantT_co], Protocol[CompliantT_co]):
     def get_categories(self) -> CompliantT_co: ...
 
 
-class DateTimeNamespace(_StoresCompliant[CompliantT_co], Protocol[CompliantT_co]):
+class DateTimeNamespace(NamespaceAccessor[CompliantT_co], Protocol[CompliantT_co]):
     _accessor: ClassVar[Accessor] = "dt"
 
     def to_string(self, format: str) -> CompliantT_co: ...
@@ -61,7 +69,7 @@ class DateTimeNamespace(_StoresCompliant[CompliantT_co], Protocol[CompliantT_co]
     def offset_by(self, by: str) -> CompliantT_co: ...
 
 
-class ListNamespace(_StoresCompliant[CompliantT_co], Protocol[CompliantT_co]):
+class ListNamespace(NamespaceAccessor[CompliantT_co], Protocol[CompliantT_co]):
     _accessor: ClassVar[Accessor] = "list"
 
     def get(self, index: int) -> CompliantT_co: ...
@@ -70,7 +78,7 @@ class ListNamespace(_StoresCompliant[CompliantT_co], Protocol[CompliantT_co]):
     def contains(self, item: NonNestedLiteral) -> CompliantT_co: ...
 
 
-class NameNamespace(_StoresCompliant[CompliantT_co], Protocol[CompliantT_co]):
+class NameNamespace(NamespaceAccessor[CompliantT_co], Protocol[CompliantT_co]):
     _accessor: ClassVar[Accessor] = "name"
 
     def keep(self) -> CompliantT_co: ...
@@ -81,7 +89,7 @@ class NameNamespace(_StoresCompliant[CompliantT_co], Protocol[CompliantT_co]):
     def to_uppercase(self) -> CompliantT_co: ...
 
 
-class StringNamespace(_StoresCompliant[CompliantT_co], Protocol[CompliantT_co]):
+class StringNamespace(NamespaceAccessor[CompliantT_co], Protocol[CompliantT_co]):
     _accessor: ClassVar[Accessor] = "str"
 
     def len_chars(self) -> CompliantT_co: ...
@@ -104,7 +112,7 @@ class StringNamespace(_StoresCompliant[CompliantT_co], Protocol[CompliantT_co]):
     def zfill(self, width: int) -> CompliantT_co: ...
 
 
-class StructNamespace(_StoresCompliant[CompliantT_co], Protocol[CompliantT_co]):
+class StructNamespace(NamespaceAccessor[CompliantT_co], Protocol[CompliantT_co]):
     _accessor: ClassVar[Accessor] = "struct"
 
     def field(self, name: str) -> CompliantT_co: ...
